requirements.txt
flask==2.3.2
python-telegram-bot==20.3
python-dotenv==1.0.0
requests==2.31.0
supabase==2.0.0
